make Use_Decomp optional.
Prior to this change, Use_Decomp for v3/4 would always be on when compiling scripts, no matter what.
This wasn't exactly all bad for S1 and SCD specifically, as they mainly had bugfixes.
Sonic 2, meanwhile, uses VARNAME in the scripts, which is exclusive to the decompilation and does not exist officially in the mobile remasters or Sonic Origins.
This can cause issues if S2 bytecode is built for either iteration, making the game fail to work.
Using the Custom Flags section of the Script Compiler, Use_Decomp can be used for the respective script type if people *really* want to compile with decomp exclusive code (as long as they use Use_Decomp (v3) and USE_DECOMP (v4) respectively.